+++ src/GRA.Data/Model/VendorCode.cs
@@ -19,5 +19,6 @@ namespace GRA.Data.Model
         public int? UserId { get; set; }
         public DateTime? OrderDate { get; set; }
         public DateTime? ShipDate { get; set; }
+        public bool? IsDonated { get; set; }
     }
 }