+++ src/GRA.Domain.Model/User.cs
@@ -77,5 +77,8 @@ namespace GRA.Domain.Model
         public bool HasPendingQuestionnaire { get; set; }
         public string VendorCode { get; set; }
         public string VendorCodeMessage { get; set; }
+        public bool NeedsToAnswerDonationQuestion { get; set; }
+        public string VendorCodeUrl { get; set; }
+        public bool? Donated { get; set; }
     }
 }