+++ src/GRA/TemplateToken.cs
@@ -1,0 +1,7 @@
+﻿namespace GRA
+{
+    public struct TemplateToken
+    {
+        public const string VendorCodeToken = "{Code}";
+    }
+}